Fixing the size for interrupt counts and adding a call to get the number of archetecture specific interrupts

+++ module/sir_internal.h
@@ -24,7 +24,7 @@
 
     // ==== Structure for partial reads ====
     struct partial_read_state{
-        unsigned int interrupts;
+        u64 interrupts;
         char ind;
         struct mutex lock;
     } ;
